@@ -1,6 +1,10 @@
-﻿namespace Deepgram.DeepgramEventArgs
+﻿namespace Deepgram.DeepgramEventArgs;
+
+public class ConnectionErrorEventArgs : EventArgs
 {
-    public class ConnectionErrorEventArgs(Exception ex) : EventArgs
+    public Exception Exception { get; set; }
+    public ConnectionErrorEventArgs(Exception ex)
     {
+        Exception = ex;
     }
 }
