@@ -2,26 +2,21 @@
 public static class MockHttpClient
 {
     public static HttpClient CreateHttpClientWithResult<T>(
-       T result, HttpStatusCode code = HttpStatusCode.OK, string? url = $"https://{Constants.DEFAULT_URI}")
+       T result, HttpStatusCode code = HttpStatusCode.OK, string url = $"https://{Constants.DEFAULT_URI}")
     {
 
-        var httpClient = new HttpClient(new MockHttpMessageHandler<T>(result, code))
+        return new HttpClient(new MockHttpMessageHandler<T>(result, code))
         {
-            BaseAddress = new Uri(url!)
+            BaseAddress = new Uri(url)
         };
-
-        return httpClient;
     }
 
     public static HttpClient CreateHttpClientWithException<T>(
        T result, HttpStatusCode code = HttpStatusCode.OK)
     {
-
-        var httpClient = new HttpClient(new MockHttpMessageHandler<T>(result, code))
+        return new HttpClient(new MockHttpMessageHandler<T>(result, code))
         {
             BaseAddress = new Uri($"https://{Constants.DEFAULT_URI}")
         };
-
-        return httpClient;
     }
 }
