@@ -3,12 +3,13 @@
 public class AbstractRestfulClientTests
 {
     IHttpClientFactory _httpClientFactory;
-    string _apiKey;
+    DeepgramClientOptions _clientOptions;
 
     [SetUp]
     public void Setup()
     {
-        _apiKey = new Faker().Random.Guid().ToString();
+
+        _clientOptions = new DeepgramClientOptions(new Faker().Random.Guid().ToString());
         _httpClientFactory = Substitute.For<IHttpClientFactory>();
     }
 
@@ -19,7 +20,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<GetProjectsResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithResult(expectedResponse, HttpStatusCode.OK);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
 
         // Act
         var result = await client.GetAsync<GetProjectsResponse>(Constants.PROJECTS);
@@ -40,7 +41,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<GetProjectsResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithResult(expectedResponse, HttpStatusCode.BadRequest);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
         // Act & Assert
         client.Invoking(y => y.GetAsync<GetProjectsResponse>(Constants.PROJECTS))
              .Should().ThrowAsync<HttpRequestException>();
@@ -57,7 +58,7 @@ public class AbstractRestfulClientTests
         expectedResponse.ProjectId = "1";
         var httpClient = MockHttpClient.CreateHttpClientWithResult(expectedResponse, HttpStatusCode.OK);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
         // Act
         var result = await client.GetAsync<GetProjectResponse>($"{Constants.PROJECTS}/1");
 
@@ -78,7 +79,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<CreateProjectKeyResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithResult(expectedResponse, HttpStatusCode.OK);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
         // Act
         var result = await client.PostAsync<CreateProjectKeyResponse>(Constants.PROJECTS, new StringContent(string.Empty));
 
@@ -98,7 +99,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<SyncPrerecordedResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithResult(expectedResponse, HttpStatusCode.OK);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
         // Act
         var result = await client.PostAsync<SyncPrerecordedResponse>(Constants.PROJECTS, new StringContent(string.Empty));
 
@@ -117,7 +118,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<SyncPrerecordedResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithResult(expectedResponse, HttpStatusCode.BadRequest);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
 
         // Act & Assert      
 
@@ -131,7 +132,7 @@ public class AbstractRestfulClientTests
         // Arrange   
         var httpClient = MockHttpClient.CreateHttpClientWithResult(new VoidResponse(), HttpStatusCode.OK);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
 
         // Act & Assert
         client.Invoking(y => y.Delete($"{Constants.PROJECTS}/1"))
@@ -145,7 +146,7 @@ public class AbstractRestfulClientTests
         // Arrange    
         var httpClient = MockHttpClient.CreateHttpClientWithException(new MessageResponse(), HttpStatusCode.BadRequest);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
 
         // Act & Assert
         _ = client.Invoking(y => y.Delete(Constants.PROJECTS))
@@ -159,7 +160,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<MessageResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithResult(expectedResponse, HttpStatusCode.OK);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory) { };
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
 
         // Act
         var result = await client.DeleteAsync<MessageResponse>($"{Constants.PROJECTS}/1");
@@ -180,7 +181,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<MessageResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithException(expectedResponse, HttpStatusCode.BadRequest);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
 
         // Act & Assert
         client.Invoking(y => y.DeleteAsync<MessageResponse>(Constants.PROJECTS))
@@ -195,7 +196,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<MessageResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithResult(expectedResponse, HttpStatusCode.OK);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory) { };
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
 
         // Act
         var result = await client.PatchAsync<MessageResponse>($"{Constants.PROJECTS}/1", new StringContent(string.Empty));
@@ -216,7 +217,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<MessageResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithException(expectedResponse, HttpStatusCode.BadRequest);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
 
         //Act & Assert
         client.Invoking(y => y.PatchAsync<MessageResponse>(Constants.PROJECTS, new StringContent(string.Empty)))
@@ -230,7 +231,7 @@ public class AbstractRestfulClientTests
         var expectedResponse = new AutoFaker<MessageResponse>().Generate();
         var httpClient = MockHttpClient.CreateHttpClientWithResult(expectedResponse, HttpStatusCode.OK);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory) { };
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
 
         // Act
         var result = await client.PutAsync<MessageResponse>($"{Constants.PROJECTS}/1", new StringContent(string.Empty));
@@ -250,7 +251,7 @@ public class AbstractRestfulClientTests
         // Arrange       
         var httpClient = MockHttpClient.CreateHttpClientWithException(new MessageResponse(), HttpStatusCode.BadRequest);
         _httpClientFactory.CreateClient(Constants.HTTPCLIENT_NAME).Returns(httpClient);
-        var client = new ConcreteRestClient(_apiKey, _httpClientFactory);
+        var client = new ConcreteRestClient(_clientOptions, _httpClientFactory);
         // Act & Assert
         client.Invoking(y => y.PutAsync<MessageResponse>(Constants.PROJECTS, new StringContent(string.Empty)))
              .Should().ThrowAsync<HttpRequestException>();
