@@ -3,11 +3,10 @@
 /// <summary>
 ///  Client containing methods for interacting with API's to manage project(s)
 /// </summary>
-/// <param name="apiKey">ApiKey used for Authentication Header and is required</param>
 /// <param name="httpClientFactory"><see cref="IHttpClientFactory"/> for creating instances of HttpClient for making Rest calls</param>
 /// <param name="deepgramClientOptions"><see cref="DeepgramClientOptions"/> for HttpClient Configuration</param>
-public class ManageClient(string? apiKey, IHttpClientFactory httpClientFactory, DeepgramClientOptions? deepgramClientOptions = null)
-    : AbstractRestClient(apiKey, httpClientFactory, deepgramClientOptions)
+public class ManageClient(DeepgramClientOptions deepgramClientOptions, IHttpClientFactory httpClientFactory)
+    : AbstractRestClient(deepgramClientOptions, httpClientFactory)
 {
     readonly string _urlPrefix = $"/{Constants.API_VERSION}/{Constants.PROJECTS}";
 
