@@ -3,11 +3,10 @@
 /// <summary>
 /// Constructor that take a IHttpClientFactory
 /// </summary>
-/// <param name="apiKey">ApiKey used for Authentication Header and is required</param> 
 /// <param name="httpClientFactory"><see cref="IHttpClientFactory"/> for creating instances of HttpClient for making Rest calls</param>
 /// <param name="deepgramClientOptions"><see cref="DeepgramClientOptions"/> for HttpClient Configuration</param>
-public class PrerecordedClient(string? apiKey, IHttpClientFactory httpClientFactory, DeepgramClientOptions? deepgramClientOptions = null)
-    : AbstractRestClient(apiKey, httpClientFactory, deepgramClientOptions)
+public class PrerecordedClient(DeepgramClientOptions deepgramClientOptions, IHttpClientFactory httpClientFactory)
+    : AbstractRestClient(deepgramClientOptions, httpClientFactory)
 {
     readonly string _urlPrefix = $"/{Constants.API_VERSION}/{Constants.LISTEN}";
     #region NoneCallBacks
